+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * add -c file for assigning lua scriptable command-line overrides
  * headless runner for arcan-net host appl can be access via ANET\_RUNNER env.
  * spawning server-side Lua runner if matching appl found, controls message routing
+ * introduce rekeying command for forward secrecy, placeholder PQ step-up and resumption
 
 ## 0.6.3
 ## Lua